+++ app/system/internal/handler/member/user/reset_pwd_handler.go
@@ -1,0 +1,31 @@
+// Code scaffolded by goctl. Safe to edit.
+// goctl 1.10.0
+
+package user
+
+import (
+	"net/http"
+
+	"github.com/zeromicro/go-zero/rest/httpx"
+	"ovra/app/system/internal/logic/member/user"
+	"ovra/app/system/internal/svc"
+	"ovra/app/system/internal/types"
+)
+
+func ResetPwdHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var req types.MemberUserResetPwdReq
+		if err := httpx.Parse(r, &req); err != nil {
+			httpx.ErrorCtx(r.Context(), w, err)
+			return
+		}
+
+		l := user.NewResetPwdLogic(r.Context(), svcCtx)
+		err := l.ResetPwd(&req)
+		if err != nil {
+			httpx.ErrorCtx(r.Context(), w, err)
+		} else {
+			httpx.OkJsonCtx(r.Context(), w, nil)
+		}
+	}
+}